test: replace util.py with conftest.py in tests
